[198] Add styles
