+++ src/App.scss
@@ -28,24 +28,28 @@
   }
 
   &__nav {
-    z-index: 1;
+    display: flex;
+    justify-content: space-around;
     background: white;
     width: 100%;
     flex: 0 0 3rem;
     margin: 0;
-    text-align: center;
     color: black;
     border-top: 2px solid black;
-    font-size: 2rem;
   }
 
   &__agile {
     border: none;
     background: white;
+    font-size: 1.5rem;
 
     &-img {
-      height: 3.75rem;
-      padding: 0.5rem;
+      height: 2rem;
+    }
+
+    p {
+      font-size: 0.7rem;
+      margin: 0;
     }
   }
 }
