+++ src/components/FreakDetails/FreakDetails.scss
@@ -1,7 +1,7 @@
 .freak-details {
   border-radius: 2rem;
   text-align: center;
-  margin: 9rem auto 0;
+  margin: 7rem auto 0;
   border: 1rem solid;
   background: rgba(16, 95, 191, 0.9);
   color: white;
