+++ src/pages/ViewFreaks/ViewFreaks.scss
